# api/openapi.yaml
openapi: 3.1.0
info:
  title: Wealthboard API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api/v1
paths:
...
  /exports/user:
    get:
      operationId: exportUserArchive
      summary: Download the current user's portable JSON archive
      description: Bearer API keys require `exports:read`; authenticated browser sessions may also export. The download contains private owner-scoped financial and estate data, excludes authentication secrets and user IDs, and must not be cached.
      security: [{ sessionCookie: [] }, { apiKey: [] }]
      responses:
        "200":
          description: Versioned Wealthboard user archive
          headers:
            Content-Disposition:
              { $ref: "#/components/headers/UserArchiveDisposition" }
            Cache-Control: { $ref: "#/components/headers/NoStore" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserArchive" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /exports/transactions.csv:
    get:
      operationId: exportTransactionsCsv
      summary: Download the current user's transactions as CSV
      description: Bearer API keys require `exports:read`; authenticated browser sessions may also export. The private owner-scoped export must not be cached. Monetary columns use integer minor units.
      security: [{ sessionCookie: [] }, { apiKey: [] }]
      responses:
        "200":
          description: UTF-8 transaction CSV
          headers:
            Content-Disposition:
              { $ref: "#/components/headers/TransactionsDisposition" }
            Cache-Control: { $ref: "#/components/headers/NoStore" }
          content:
            text/csv:
              schema: { type: string, contentMediaType: "text/csv" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /exports/accounts.csv:
    get:
      operationId: exportAccountsCsv
      summary: Download the current user's accounts as CSV
      description: Bearer API keys require `exports:read`; authenticated browser sessions may also export. The private owner-scoped export must not be cached. Monetary columns use integer minor units.
      security: [{ sessionCookie: [] }, { apiKey: [] }]
      responses:
        "200":
          description: UTF-8 account CSV
          headers:
            Content-Disposition:
              { $ref: "#/components/headers/AccountsDisposition" }
            Cache-Control: { $ref: "#/components/headers/NoStore" }
          content:
            text/csv:
              schema: { type: string, contentMediaType: "text/csv" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /restore/user:
    post:
      operationId: restoreUserArchive
      summary: Replace the current user's portable data from an archive
      description: Requires a browser session, trusted Origin, and `X-CSRF-Token`; Bearer API keys are not accepted. Accepts a raw JSON archive or one multipart `file`. Archive content is limited to 25 MiB; the raw HTTP body permits at most 1 MiB additional multipart overhead. Restore validates owner-scoped relationships and never restores authentication secrets or another user's identity.
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/MutationCSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserArchive" }
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Versioned JSON archive, maximum 25 MiB.
      responses:
        "200":
          description: Counts of restored records
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RestoreSummary" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: wealthboard_session
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: wbk_v1
      description: API-key scopes are enforced per operation as documented in each operation description.
  parameters:
...
    MutationCSRFToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required when authenticating with `sessionCookie`; ignored for Bearer API keys.
      schema:
        type: string
        minLength: 1
...
  headers:
    NoStore:
      description: Prevents storage of private financial or credential metadata.
      schema:
        type: string
        const: no-store
    UserArchiveDisposition:
      description: Attachment filename for the JSON user archive.
      schema:
        type: string
        const: 'attachment; filename="wealthboard-export.json"'
    TransactionsDisposition:
      description: Attachment filename for the transaction CSV.
      schema:
        type: string
        const: 'attachment; filename="wealthboard-transactions.csv"'
    AccountsDisposition:
      description: Attachment filename for the account CSV.
      schema:
        type: string
        const: 'attachment; filename="wealthboard-accounts.csv"'
...
  schemas:
...
    PortableRecord:
      type: object
      description: One owner-scoped archive row. Database `*_minor` fields are integer minor units; rates, prices, quantities, and ratios are exact decimal strings. User ownership columns and authentication secrets are excluded.
      additionalProperties: true
    UserArchive:
      type: object
      additionalProperties: false
      required:
        [
          format,
          version,
          exportedAt,
          settings,
          categories,
          institutions,
          accounts,
          transactions,
          valuations,
          exchangeRates,
          goals,
          goalContributionPlans,
          goalMilestones,
          goalAlertDismissals,
          beneficiaries,
          estatePlans,
          estateAccountDirectives,
          estateAllocations,
          estateResiduaryAllocations,
          estatePlanSnapshots,
          investmentInstruments,
          positionEvents,
          securityPrices,
          positionReconciliations,
          accountConversions,
        ]
      properties:
        format: { type: string, const: wealthboard-user-export }
        version: { type: integer, const: 8 }
        exportedAt: { type: string, format: date-time }
        settings:
          {
            type: object,
            additionalProperties: true,
            description: Owner settings without credentials or session data.,
          }
        categories:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        institutions:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        accounts:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        transactions:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        valuations:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        exchangeRates:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        goals:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        goalContributionPlans:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        goalMilestones:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        goalAlertDismissals:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        beneficiaries:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        estatePlans:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        estateAccountDirectives:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        estateAllocations:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        estateResiduaryAllocations:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        estatePlanSnapshots:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        investmentInstruments:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        positionEvents:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        securityPrices:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        positionReconciliations:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
        accountConversions:
          {
            type: array,
            items: { $ref: "#/components/schemas/PortableRecord" },
          }
    RestoreSummary:
      type: object
      additionalProperties: false
      required:
        [
          institutions,
          accounts,
          transactions,
          goals,
          milestones,
          beneficiaries,
          estatePlans,
          estateSnapshots,
          investmentInstruments,
          positionEvents,
          securityPrices,
          positionReconciliations,
        ]
      properties:
        institutions: { type: integer, minimum: 0 }
        accounts: { type: integer, minimum: 0 }
        transactions: { type: integer, minimum: 0 }
        goals: { type: integer, minimum: 0 }
        milestones: { type: integer, minimum: 0 }
        beneficiaries: { type: integer, minimum: 0 }
        estatePlans: { type: integer, minimum: 0 }
        estateSnapshots: { type: integer, minimum: 0 }
        investmentInstruments: { type: integer, minimum: 0 }
        positionEvents: { type: integer, minimum: 0 }
        securityPrices: { type: integer, minimum: 0 }
        positionReconciliations: { type: integer, minimum: 0 }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
...
  responses:
...
    Problem:
      description: Request failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"